pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lantern"
version = "0.1.0"
description = "Lantern governed workflow runtime"
requires-python = ">=3.11"
dependencies = ["pyyaml", "jsonschema"]

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
